feat: add state declarations

+++ src/networkLayerPlus/ndnRouting/dataPack/PacketCommon.h
@@ -4,7 +4,8 @@
 #include "util/util.h"
 #include "util/json/json.hpp"
 #include "util/printable/Printable.h"
-typedef int RouterID;
+//typedef int RouterID;
+typedef uint64_t RouterID;
 
 enum LinkStateType { ADJ = 0, RCH = 1 };
 inline std::string getNameForLinkStateType(LinkStateType t) {
